requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
apscheduler==3.10.4
aiohttp==3.9.3
python-dotenv==1.0.0
psycopg2-binary==2.9.8